Initial template for DragonFly 5.0
Manual install with these boot_commands keystrokes works fine.
But something about doing it through packer is causing a kernel panic.

All of the partitions, fs creation, installation and boot block steps happen.
Then "Configure the system" is selected, the config menu appears for a second
and panic!

Commiting this "should be working" version and will explore fixes or workarounds.